+++ netsgiro/__init__.py
@@ -8,8 +8,4 @@ from netsgiro.objects import *  # noqa: Reexport
 
 from netsgiro import constants, enums, objects
 
-__all__ = (
-    constants.__all__ +
-    enums.__all__ +
-    objects.__all__
-)
+__all__ = constants.__all__ + enums.__all__ + objects.__all__