+++ netsgiro/constants.py
@@ -1,8 +1,6 @@
 """Constants that are useful when working with OCR files."""
 
-__all__ = [
-    'NETS_ID'
-]
+__all__ = ['NETS_ID']
 
 
 #: Nets' data sender/recipient ID.