+++ netsgiro/validators.py
@@ -10,8 +10,11 @@ def str_of_length(length):
         instance_of(str)(instance, attribute, value)
         if len(value) != length:
             raise ValueError(
-                '{0.name} must be exactly {1} chars, got {2!r}'
-                .format(attribute, length, value))
+                '{0.name} must be exactly {1} chars, got {2!r}'.format(
+                    attribute, length, value
+                )
+            )
+
     return validator
 
 
@@ -23,6 +26,9 @@ def str_of_max_length(length):
         if len(value) > length:
             raise ValueError(
                 '{0.name} must be at most {1} chars, '
-                'got {2!r} which is {3} chars'
-                .format(attribute, length, value, len(value)))
+                'got {2!r} which is {3} chars'.format(
+                    attribute, length, value, len(value)
+                )
+            )
+
     return validator